more automatic build stuff

+++ openwrt/install.sh
@@ -33,6 +33,14 @@ if [ ! -e kernel ]; then
   mkdir kernel;
 fi
 
+if [ ! -e testbed-server ]; then
+  mkdir testbed-server;
+fi
+
+if [ ! -e testbed-server/srv/boot ]; then
+  mkdir -p testbed-server/srv/boot;
+fi
+
 . ./config/config.common
 
 if [ "x$CONFIG" != "x" ]; then
@@ -40,7 +48,6 @@ if [ "x$CONFIG" != "x" ]; then
     . $CONFIG
   fi
 fi
-
 
 for i in $EXTRA_FEEDS; do
   (cd feeds; git clone $i)
@@ -53,7 +60,8 @@ for i in $BUILD_ARCH; do
 
   OPENWRT_PATH=$OPENWRT_VERSION\-$i
 
-#if [ "x$DIR" = "y" ]; then
+if [ "x$DIR" = "y" ]; then
+
   (git clone $OPENWRT_GITURL $OPENWRT_PATH)
   (cd $OPENWRT_PATH; git reset --hard $OPENWRT_REVISION)
 
@@ -73,7 +81,9 @@ for i in $BUILD_ARCH; do
 
   (cd $OPENWRT_PATH; yes "" | make oldconfig; make V=99 -j10) | tee $OPENWRT_PATH/openwrt_build.log
 
-  echo "export OPENWRT_PATH=$OPENWRT_PATH" > $OPENWRT_PATH/toolchain.bashrc
+fi
+
+  echo "export OPENWRT_PATH=$DIR/$OPENWRT_PATH" > $OPENWRT_PATH/toolchain.bashrc
 
   TOOLCHAIN_DIR=`(cd $OPENWRT_PATH; for i in staging_dir/tool*; do if [ -e $i/bin/ ]; then echo $i; break; fi; done)`
 
@@ -83,7 +93,15 @@ for i in $BUILD_ARCH; do
   echo "export PATH=\$PATH:\$TOOLCHAIN_DIR/bin/" >> $OPENWRT_PATH/toolchain.bashrc
   echo "export PATH=\$PATH:\$STAGING_DIR/host/bin" >> $OPENWRT_PATH/toolchain.bashrc
 
-#fi
+  OPENWRT_ROOTFS_DIR=`(cd $OPENWRT_PATH; for i in staging_dir/target-*; do if [ -e $i/bin/ ]; then echo $i; break; fi; done)`
+
+  echo "export OPENWRT_ROOTFS_DIR=\$OPENWRT_PATH/$OPENWRT_ROOTFS_DIR" >> $OPENWRT_PATH/toolchain.bashrc
+
+  . $OPENWRT_PATH/toolchain.bashrc
+
+  if [ ! -e $DIR/testbed-server/srv/nfsroot-$i ]; then
+      (cd $OPENWRT_ROOTFS_DIR; cp -r root-* $DIR/testbed-server/srv/nfsroot-$i)
+  fi
 
   if [ "x$KERNEL_CONFIG" != "x" ]; then
 
@@ -91,13 +109,47 @@ for i in $BUILD_ARCH; do
     LINUXTOOL_DIR=`cd $OPENWRT_PATH/$BUILD_DIR; ls -d linux-*`
     LINUX_DIR=`cd $OPENWRT_PATH/$BUILD_DIR/$LINUXTOOL_DIR; ls -d linux-3*`
 
-    cp -r $OPENWRT_PATH/$BUILD_DIR/$LINUXTOOL_DIR/$LINUX_DIR kernel/$LINUX_DIR\-$i
+    if [ ! -e kernel/$LINUX_DIR\-$i ]; then
+        cp -r $OPENWRT_PATH/$BUILD_DIR/$LINUXTOOL_DIR/$LINUX_DIR kernel/$LINUX_DIR\-$i
+    fi
     cp config/$KERNEL_CONFIG kernel/$LINUX_DIR\-$i/
     cp config/$KERNEL_CONFIG kernel/$LINUX_DIR\-$i/.config
 
     cp tools/*.sh kernel/$LINUX_DIR\-$i/
 
-    
+    if [ "x$i" = "xx86" ]; then
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i CPUS=1 sh ./kernelbuild.sh clean )
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i CPUS=60 sh ./kernelbuild.sh bzImage modules)
+
+        rm -rf $DIR/testbed-server/srv/nfsroot-$i/lib/modules
+        mkdir $DIR/testbed-server/srv/nfsroot-$i/lib/modules
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i INSTALL_MOD_PATH=$DIR/testbed-server/srv/nfsroot-$i/ sh ./kernelbuild.sh modules_install)
+
+        cp kernel/$LINUX_DIR\-$i/arch/x86/boot/bzImage $DIR/testbed-server/srv/boot/vmlinuz-$i
+    fi
+    if [ "x$i" = "xmips" ]; then
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i CPUS=1 sh ./kernelbuild.sh clean )
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i CPUS=60 sh ./kernelbuild.sh vmlinux modules)
+
+        rm -rf $DIR/testbed-server/srv/nfsroot-$i/lib/modules
+        mkdir $DIR/testbed-server/srv/nfsroot-$i/lib/modules
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i INSTALL_MOD_PATH=$DIR/testbed-server/srv/nfsroot-$i/ sh ./kernelbuild.sh modules_install)
+
+        cp kernel/$LINUX_DIR\-$i/vmlinux $DIR/testbed-server/srv/boot/vmlinux-$i
+    fi
+    if [ "x$i" = "xwndr3700" ]; then
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i CPUS=1 sh ./kernelbuild.sh clean )
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i CPUS=60 sh ./kernelbuild.sh vmlinux modules; sh build-tftp.sh)
+
+        rm -rf $DIR/testbed-server/srv/nfsroot-$i/lib/modules
+        mkdir $DIR/testbed-server/srv/nfsroot-$i/lib/modules
+        ( cd kernel/$LINUX_DIR\-$i/; . $OPENWRT_PATH/toolchain.bashrc; BUILD_ARCH=$i INSTALL_MOD_PATH=$DIR/testbed-server/srv/nfsroot-$i/ sh ./kernelbuild.sh modules_install)
+
+        cp kernel/$LINUX_DIR\-$i/vmlinux-wndr3700v1.uImage  $DIR/testbed-server/srv/boot/vmlinuz-$i-v210
+        cp kernel/$LINUX_DIR\-$i/vmlinux-wndr3700v2.uImage  $DIR/testbed-server/srv/boot/vmlinuz-$i
+    fi
+
+    ( cd $DIR/testbed-server/srv/boot/; chmod 644 * )
 
   fi
